#18 - Update the PostService, PostServiceImplementation, and PostController to accept valid token

+++ src/main/java/com/abichuela/b142/s02/s02app/services/PostServiceImplementation.java
@@ -1,10 +1,13 @@
 package com.abichuela.b142.s02.s02app.services;
 
+import com.abichuela.b142.s02.s02app.configs.JwtToken;
 import com.abichuela.b142.s02.s02app.models.Post;
 import com.abichuela.b142.s02.s02app.models.User;
 import com.abichuela.b142.s02.s02app.repositories.PostRepository;
 import com.abichuela.b142.s02.s02app.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,9 +16,11 @@ public class PostServiceImplementation implements PostService {
     private PostRepository postRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    JwtToken jwtToken;
 
-    public void createPost(Post newPost, Long userId) {
-        User author = userRepository.findById(userId).get();
+    public void createPost(Post newPost, String token) {
+        User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(token));
         Post post = new Post();
         post.setTitle(newPost.getTitle());
         post.setContent(newPost.getContent());
@@ -23,25 +28,38 @@ public class PostServiceImplementation implements PostService {
         postRepository.save(post);
     }
 
-    public void updatePost(Long userId, Long postId, Post updatedPost) {
+    public ResponseEntity updatePost(Long postId, Post updatedPost, String token) {
         Post existingPost = postRepository.findById(postId).get();
-        existingPost.setTitle(updatedPost.getTitle());
-        existingPost.setContent(updatedPost.getContent());
-        User newAuthor = userRepository.findById(userId).get();
-        existingPost.setUser(newAuthor);
-        postRepository.save(existingPost);
+        String authorName = existingPost.getUser().getUsername();
+        String authUsername = jwtToken.getUsernameFromToken(token);
+        if (authUsername.equals(authorName)) {
+            existingPost.setTitle(updatedPost.getTitle());
+            existingPost.setContent(updatedPost.getContent());
+            postRepository.save(existingPost);
+            return new ResponseEntity<>("Post updated successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("You are not authorized to edit this post.", HttpStatus.UNAUTHORIZED);
+        }
     }
 
-    public void deletePost(Long postId) {
-        postRepository.deleteById(postId);
+    public ResponseEntity deletePost(Long postId, String token) {
+        Post existingPost = postRepository.findById(postId).get();
+        String authorName = existingPost.getUser().getUsername();
+        String authUsername = jwtToken.getUsernameFromToken(token);
+        if (authUsername.equals(authorName)) {
+            postRepository.deleteById(postId);
+            return new ResponseEntity<>("Post deleted successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("You are not authorized to delete this post.", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     public Iterable<Post> getPosts() {
         return postRepository.findAll();
     }
 
-    public Iterable<Post> getMyPost(Long userId) {
-        User author = userRepository.findById(userId).get();
+    public Iterable<Post> getMyPosts(String token) {
+        User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(token));
         return author.getPosts();
     }
 }
